move goal enum defs to types file
