+++ src/types/goals.d.ts
@@ -1,0 +1,20 @@
+/**
+ * All the types of goals settable.
+ */
+export enum GoalTypes {
+  ITEMS = 'items',
+  LINES = 'lines',
+  MINUTES = 'minutes',
+  PAGES = 'pages',
+  WORDS = 'words',
+}
+
+/**
+ * All possible durations of goals.
+ */
+export enum GoalDurations {
+  DAILY = 'daily',
+  MONTHLY = 'monthly',
+  WEEKLY = 'weekly',
+  YEARLY = 'yearly'
+}
